refactor: Refactor metrics to a separate module
